Camera is in above icon for first toying with ReactNative
