+++ mobile/package.json
@@ -10,6 +10,7 @@
     "axios": "^0.18.0",
     "expo": "^32.0.0",
     "immutable": "^4.0.0-rc.12",
+    "native-base": "^2.12.1",
     "react": "16.5.0",
     "react-draggable": "^3.3.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
